requirements.txt
boto3
tiktoken
requests
line-bot-sdk